requirements.txt
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-language==1.3.0
ipython==7.11.1
jsonpickle==1.2
jupyter-client==6.0.0
jupyter-core==4.6.3
matplotlib==3.1.2
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
pandas==0.25.3
requests==2.22.0
requests-oauthlib==1.3.0
sanpy==0.7.11
scikit-learn==0.22.1
scipy==1.4.1
searchtweets==1.7.4
statsmodels==0.10.2
tweepy==3.8.0
tweet-parser==1.13.2